add StrategyOperationsScreen
