+++ src/components/screen.rs
@@ -11,6 +11,7 @@ mod strategies;
 mod strategy_create;
 mod strategy_confirm;
 mod strategy_contracts;
+mod strategy_operations;
 
 pub(crate) use contract::ContractScreen;
 pub(crate) use contract::ContractScreenCommands;
@@ -37,8 +38,10 @@ pub(crate) use strategy_create::CreateStrategyScreen;
 pub(crate) use strategy_create::CreateStrategyScreenCommands;
 pub(crate) use strategy_create::RenameStrategyStruct;
 pub(crate) use strategy_create::LoadStrategyStruct;
-pub(crate) use strategy_create::SelectOperationTypeStruct;
-pub(crate) use strategy_create::DocumentStruct;
+pub(crate) use strategy_operations::SelectOperationTypeStruct;
+pub(crate) use strategy_operations::DocumentStruct;
+pub(crate) use strategy_operations::StrategyOperationsScreen;
+pub(crate) use strategy_operations::StrategyOperationsScreenCommands;
 pub(crate) use strategy_confirm::ConfirmStrategyScreen;
 pub(crate) use strategy_confirm::ConfirmStrategyScreenCommands;
 pub(crate) use strategy_contracts::AddContractStruct;
